+++ src/main.cpp
@@ -10,7 +10,7 @@ int main(int argc, char* argv[])
     QCoreApplication::setOrganizationName(QStringLiteral("anSofts"));
     QCoreApplication::setOrganizationDomain(QStringLiteral("ansofts.it"));
     QCoreApplication::setApplicationName(QStringLiteral("ChronoLab"));
-    QCoreApplication::setApplicationVersion(QStringLiteral("0.1.0"));
+    QCoreApplication::setApplicationVersion(QStringLiteral("0.2.0"));
 
     QFont font(QStringLiteral("Segoe UI"));
     font.setStyleHint(QFont::SansSerif);